11627 - moved search bar to list checkbox component
