+++ src/js/components/sharedComponents/checkbox/ListCheckbox.jsx
@@ -3,11 +3,12 @@
  * Created by Josue Aguilar on 09/20/2024.
  */
 
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 
 import ListCheckboxPrimary from "./ListCheckboxPrimary";
 import SubmitHint from "../filterSidebar/SubmitHint";
+import { EntityDropdownAutocomplete } from "../../search/filters/location/EntityDropdownAutocomplete";
 
 const propTypes = {
     filters: PropTypes.object,
@@ -23,7 +24,43 @@ const defaultProps = {
 const ListCheckbox = ({
     filters, filterCategoryMapping, selectedFilters, singleFilterChange
 }) => {
-    const checkboxCategories = filterCategoryMapping.map((category) => (
+    const [searchString, setSearchString] = useState('');
+    const [filterCategory, setFilterCategory] = useState(filterCategoryMapping);
+    const [noResults, setNoResults] = useState(false);
+
+    const handleTextInputChange = (e) => {
+        setSearchString(e.target.value);
+    };
+
+    const onClear = () => {
+        setSearchString('');
+    };
+
+    const searchCategoryMapping = () => {
+        // filter out definitions based on search text
+        // eslint-disable-next-line no-unused-vars
+        const filteredDefinitions = Object.fromEntries(Object.entries(filters).filter(([key, value]) => value.toLowerCase().includes(searchString.toLowerCase())));
+
+        // filter out type mapping filters based on filteredDefinitions
+        const filteredFilters = filterCategoryMapping.map((type) => ({
+            ...type,
+            filters: type.filters.filter((v) => Object.keys(filteredDefinitions).includes(v))
+        }));
+
+        // remove any categories that do not have any filters left
+        const filteredCategories = filteredFilters.filter((type) => type.filters.length > 0);
+
+        if (filteredCategories.length > 0) {
+            setNoResults(false);
+        }
+        else {
+            setNoResults(true);
+        }
+
+        setFilterCategory(filteredCategories);
+    };
+
+    const checkboxCategories = filterCategory.map((category) => (
         <div className="checkbox-filter__wrapper" key={category.id}>
             <div
                 className="checkbox-filter__header list-checkbox"
@@ -45,10 +82,31 @@ const ListCheckbox = ({
         </div>)
     );
 
+    useEffect(() => {
+        searchCategoryMapping();
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [searchString]);
+
     return (
-        <div className="filter-item-wrap">
-            {checkboxCategories}
-            <SubmitHint selectedFilters={selectedFilters} />
+        <div className="extent-competed-filter">
+            <EntityDropdownAutocomplete
+                placeholder="Search filters..."
+                searchString={searchString}
+                enabled
+                handleTextInputChange={handleTextInputChange}
+                context={{}}
+                loading={false}
+                isClearable
+                onClear={onClear}
+                searchIcon />
+            {noResults ?
+                <div className="no-results">No results found.</div>
+                :
+                <div className="filter-item-wrap">
+                    {checkboxCategories}
+                    <SubmitHint selectedFilters={selectedFilters} />
+                </div>
+            }
         </div>
     );
 };
